info_density: Move timestamp calculations to better location.
