+++ web/src/information_density.ts
@@ -1,5 +1,6 @@
 import $ from "jquery";
 
+import {stringify_time} from "./timerender";
 import {user_settings} from "./user_settings";
 
 // These are all relative-unit values for Source Sans Pro VF,
@@ -95,6 +96,43 @@ export function set_base_typography_css_variables(): void {
     set_vertical_alignment_values(line_height_unitless);
 }
 
+export function calculate_timestamp_widths(): void {
+    const $temp_time_div = $("<div>");
+    $temp_time_div.attr("id", "calculated-timestamp-widths");
+    // Size the div to the width of the largest timestamp,
+    // but the div out of the document flow with absolute positioning.
+    $temp_time_div.css({
+        width: "max-content",
+        visibility: "hidden",
+        position: "absolute",
+        top: "-100vh",
+    });
+    // We should get a reasonable max-width by looking only at
+    // the first and last minutes of AM and PM
+    const candidate_times = ["00:00", "11:59", "12:00", "23:59"];
+
+    for (const time of candidate_times) {
+        const $temp_time_element = $("<a>");
+        $temp_time_element.attr("class", "message-time");
+        // stringify_time only returns the time, so the date here is
+        // arbitrary and only required for creating a Date object
+        const candidate_timestamp = stringify_time(Date.parse(`1999-07-01T${time}`));
+        $temp_time_element.text(candidate_timestamp);
+        $temp_time_div.append($temp_time_element);
+    }
+
+    // Append the <div> element to calculate the maximum rendered width
+    $("body").append($temp_time_div);
+    const max_timestamp_width = $temp_time_div.width();
+    // Set the width as a CSS variable
+    $(":root").css("--message-box-timestamp-column-width", `${max_timestamp_width}px`);
+    // Clean up by removing the temporary <div> element
+    $temp_time_div.remove();
+}
+
 export function initialize(): void {
     set_base_typography_css_variables();
+    // We calculate the widths of a candidate set of timestamps,
+    // and use the largest to set `--message-box-timestamp-column-width`
+    calculate_timestamp_widths();
 }
