+++ web/src/server_events_dispatch.js
@@ -751,7 +751,7 @@ export function dispatch_normal_event(event) {
             }
             if (event.property === "twenty_four_hour_time") {
                 // Recalculate timestamp column width
-                message_lists.calculate_timestamp_widths();
+                information_density.calculate_timestamp_widths();
                 // Rerender the whole message list UI
                 for (const msg_list of message_lists.all_rendered_message_lists()) {
                     msg_list.rerender();
